@bollhals 2021-02-11
Nice improvements! Looks very much worth it.

Some of your improvements I also applied in #962, so we should look how we coordinate these two PRs to get the maximum out of it 🔥 🚀 

Out of curiosity, while changing it, what change did bring the most benefit? Is shifting & then write as ulong that much faster?