@stebet 2021-02-11
The UInt64 change helped get rid of method calls and calls to Span.Slice, which even if they are cheap, add up quickly on a hot-path.

The biggest factor was splitting up `OutgoingCommand` to reduce code branches (ifs) and extra method calls.

I have an idea for a few more optimizations which I'll try out tomorrow.