@stebet 2021-02-12
Updated with latest perf numbers