@bollhals 2021-02-18
Go for it. I'll update #962 afterwards to pickup these improvements and extend where possible.